+++ sdk/provisioning/Azure.Provisioning/src/BicepValueExtensions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
+using Azure.Provisioning.Expressions;
 using Azure.Provisioning.Primitives;
 
 namespace Azure.Provisioning;
@@ -34,4 +34,28 @@ public static class BicepValueExtensions
     }
 
     // TODO: Add more common casts
+
+    /// <summary>
+    /// Convert a IBicepValue into a BicepExpression by its reference to represent its hierarchy.
+    /// </summary>
+    /// <param name="bicepValue"></param>
+    /// <returns></returns>
+    public static BicepExpression ToBicepExpression(this IBicepValue bicepValue)
+    {
+        // if self is set, we could build an expression as a reference of this member
+        if (bicepValue.Self is not null)
+        {
+            return bicepValue.Self.GetReference();
+        }
+        // if self is not set, but the value of this is an expression, we return that expression
+        else if (bicepValue.Kind == BicepValueKind.Expression)
+        {
+            return bicepValue.Expression ?? BicepSyntax.Null();
+        }
+        // otherwise, we return whatever this compiles into
+        else
+        {
+            return bicepValue.Compile();
+        }
+    }
 }